Rename res to resource
