+++ connectors/src/api/create_connector.ts
@@ -24,7 +24,7 @@ import type {
 import { errorFromAny } from "@connectors/lib/error";
 import logger from "@connectors/logger/logger";
 import { apiError, withLogging } from "@connectors/logger/withlogging";
-import { ConnectorResource } from "@connectors/resources/connector_res";
+import { ConnectorResource } from "@connectors/resources/connector_resource";
 
 type ConnectorCreateResBody = WithConnectorsAPIErrorReponse<ConnectorType>;
