+++ connectors/src/api/set_connector_permissions.ts
@@ -6,7 +6,7 @@ import * as reporter from "io-ts-reporters";
 
 import { SET_CONNECTOR_PERMISSIONS_BY_TYPE } from "@connectors/connectors";
 import { apiError, withLogging } from "@connectors/logger/withlogging";
-import { ConnectorResource } from "@connectors/resources/connector_res";
+import { ConnectorResource } from "@connectors/resources/connector_resource";
 
 type SetConnectorPermissionsRes = WithConnectorsAPIErrorReponse<{
   success: true;
