+++ connectors/src/connectors/intercom/lib/conversation_permissions.ts
@@ -12,7 +12,7 @@ import {
 import type { ConnectorPermissionRetriever } from "@connectors/connectors/interface";
 import { IntercomTeam } from "@connectors/lib/models/intercom";
 import logger from "@connectors/logger/logger";
-import { ConnectorResource } from "@connectors/resources/connector_res";
+import { ConnectorResource } from "@connectors/resources/connector_resource";
 
 export async function allowSyncTeam({
   connector,
