+++ connectors/src/connectors/intercom/temporal/activities.ts
@@ -28,7 +28,7 @@ import {
 } from "@connectors/lib/models/intercom";
 import { syncStarted, syncSucceeded } from "@connectors/lib/sync_status";
 import logger from "@connectors/logger/logger";
-import { ConnectorResource } from "@connectors/resources/connector_res";
+import { ConnectorResource } from "@connectors/resources/connector_resource";
 
 const INTERCOM_CONVO_BATCH_SIZE = 20;
