+++ connectors/src/connectors/webcrawler/temporal/activities.ts
@@ -29,7 +29,7 @@ import {
   syncSucceeded,
 } from "@connectors/lib/sync_status";
 import logger from "@connectors/logger/logger";
-import { ConnectorResource } from "@connectors/resources/connector_res";
+import { ConnectorResource } from "@connectors/resources/connector_resource";
 
 const CONCURRENCY = 4;
