+++ connectors/src/resources/base_resource.ts
