+++ connectors/src/resources/connector_resource.ts
